+++ src/identity.ts
@@ -1,4 +1,0 @@
-/** Just return the same value. */
-export function identity<T>(value: T): T {
-  return value;
-}