+++ src/promise.ts
@@ -1,9 +1,9 @@
+import { isError } from './is_error';
+
 /**
  * Creates new resolved promise if value is not an error,
  * otherwire returns rejected promise.
  */
 export function promise<T>(value: T | PromiseLike<T>): Promise<T> {
-  return value instanceof Error
-    ? Promise.reject(value)
-    : Promise.resolve(value);
+  return isError(value) ? Promise.reject(value) : Promise.resolve(value);
 }